Level 4: Tunnel Vision
Made a new level designed to be easier than my previous one! It's named level 4 for now, but I think this might be the new level 3 so that the difficulty curve is more even! And the previous level 3 "Twinkle" will likely either be used as one of the final levels of the 2nd build or be redesigned entirely! Depends on when we're able to have the player jump through present platforms!